🧹 [Refactor observability module]
🎯 What: Extracted the monolithic `observability.py` file (over 800 lines) into a properly structured Python package module containing smaller, purpose-driven files (`metrics.py`, `tracing.py`, `logging.py`, `profiling.py`, `health.py`, `decorators.py`, and `context.py`).
💡 Why: Files over 500 lines contain multiple responsibilities. Breaking this down improves maintainability, organization, and readability, aligning with the Single Responsibility Principle, without altering existing functionality.
✅ Verification: Ran `pytest` in the `packages/workhorse` directory to ensure all tests pass and functionality is unbroken. Ran `ruff check` to enforce linting standards.
✨ Result: The monolith is now modular, enabling easier navigation, extension, and testing of observability features.
